Camera

Cameras.Camera

Cameras.Camera

Cameras.calibrate!

Cameras.calibrate!

Cameras.calibrate_extrinsic!

Cameras.calibrate_intrinsic!

Cameras.compute_homogeneous_matrix

Cameras.projection_matrix

Cameras.rq

Cameras.Camera — Type

Camera()
Camera{T}()

Construct Camera object.

Cameras.Camera — Method

camera(X)

Calculate coordinates in image from actual coordinates X. camera should be calibrate!d before using this function.

Cameras.calibrate! — Method

calibrate!(camera::Camera, chessboards::Vector{<: Chessboard}; [gridspace = 1])

Calibrate camera from chessboards.

See also calibrate_intrinsic! and calibrate_extrinsic!.

Cameras.calibrate! — Method

calibrate!(camera::Camera, xᵢ => Xᵢ)

Calibrate camera from the pair of coordinates of image xᵢ and its corresponding actual coordinates Xᵢ. The elements of xᵢ should be vector of length 2 and those of Xᵢ should be vector of length 3.

Cameras.calibrate_extrinsic! — Method

calibrate!(camera::Camera, chessboard::Chessboard; [gridspace = 1])

Calibrate extrinsic parameters of camera from chessboard.

Cameras.calibrate_intrinsic! — Method

calibrate!(camera::Camera, chessboards::Vector{<: Chessboard})

Calibrate intrinsic parameters of camera from chessboards.

Cameras.compute_homogeneous_matrix — Method

compute_homogeneous_matrix(xᵢ => Xᵢ)

Compute H in $\bm{x} \simeq \bm{H} \bm{X}$.

Cameras.projection_matrix — Method

projection_matrix(camera)

Return projection matrix to convert actual coordinate to corresponding 2D image coordinates.

Cameras.rq — Method

rq(A)

Compute the RQ factorization of the matrix A.