Digital Image Correlation (DIC)

Cameras.coarse_search

Cameras.fine_search

Cameras.neighborindices

Cameras.walkindices

Cameras.zncc

Cameras.coarse_search — Method

coarse_search(subset, image; region = CartesianIndices(image)) -> indices, C

Perform coarse search subset in image using DIC. Return the indices which has the highest correlation with subset. Use image[indices] to get the found part of image. The searching region (entire image by default) can also be specified by CartesianIndices to reduce computations.

See also neighborindices.

Examples

julia> image = rand(10,10);

julia> subset = image[3:5, 2:3];

julia> coarse_search(subset, image)
(CartesianIndex{2}[CartesianIndex(3, 2) CartesianIndex(3, 3); CartesianIndex(4, 2) CartesianIndex(4, 3); CartesianIndex(5, 2) CartesianIndex(5, 3)], 1.0)

Cameras.fine_search — Method

fine_search(subset, image, first_guess::PixelIndices) -> center, C

Perform fine search subset in image based on the Newton-Raphson method. The results by coarse_search can be used as first_guess. Note that returned center is a center coordinates (not integer any more) of searched subset in image.

Examples

julia> image = Cameras.testimage("buffalo");

julia> subset = image[100:300, 300:500];

julia> center, C = fine_search(subset, image, CartesianIndices((101:301, 301:501)))
([200.00000782067005, 400.00001094427904], 0.9999999999437896)

Cameras.neighborindices — Method

Cameras.neighborindices(subset::PixelIndices, image, npixels::Int)

Return npixels outer indices around subset. Violated indices in image are cut automatically. This is useful to give region in coarse_search.

julia> image = rand(10,10);

julia> Cameras.neighborindices(CartesianIndices((4:6, 3:6)), image, 2)
7×8 CartesianIndices{2, Tuple{UnitRange{Int64}, UnitRange{Int64}}}:
 CartesianIndex(2, 1)  CartesianIndex(2, 2)  …  CartesianIndex(2, 8)
 CartesianIndex(3, 1)  CartesianIndex(3, 2)     CartesianIndex(3, 8)
 CartesianIndex(4, 1)  CartesianIndex(4, 2)     CartesianIndex(4, 8)
 CartesianIndex(5, 1)  CartesianIndex(5, 2)     CartesianIndex(5, 8)
 CartesianIndex(6, 1)  CartesianIndex(6, 2)     CartesianIndex(6, 8)
 CartesianIndex(7, 1)  CartesianIndex(7, 2)  …  CartesianIndex(7, 8)
 CartesianIndex(8, 1)  CartesianIndex(8, 2)     CartesianIndex(8, 8)

Cameras.walkindices — Method

Cameras.walkindices(subset, image; region = CartesianIndices(image))

Return indices to walk image with size of subset.

julia> image = rand(4,4);

julia> subset = rand(2,2);

julia> Cameras.walkindices(subset, image)
3×3 Matrix{CartesianIndices{2, Tuple{UnitRange{Int64}, UnitRange{Int64}}}}:
 [CartesianIndex(1, 1) CartesianIndex(1, 2); CartesianIndex(2, 1) CartesianIndex(2, 2)]  …  [CartesianIndex(1, 3) CartesianIndex(1, 4); CartesianIndex(2, 3) CartesianIndex(2, 4)]
 [CartesianIndex(2, 1) CartesianIndex(2, 2); CartesianIndex(3, 1) CartesianIndex(3, 2)]     [CartesianIndex(2, 3) CartesianIndex(2, 4); CartesianIndex(3, 3) CartesianIndex(3, 4)]
 [CartesianIndex(3, 1) CartesianIndex(3, 2); CartesianIndex(4, 1) CartesianIndex(4, 2)]     [CartesianIndex(3, 3) CartesianIndex(3, 4); CartesianIndex(4, 3) CartesianIndex(4, 4)]

Cameras.zncc — Method

zncc(image1, image2)

Perform zero-mean normalized cross-correlation between two images.

\[C = \frac{\sum{(A_{ij} - \bar{A}_{ij}) (B_{ij} - \bar{B}_{ij})}}{\sqrt{\sum{(A_{ij} - \bar{A}_{ij})^2} \sum{(B_{ij} - \bar{B}_{ij})^2}}}\]